refactor(server): split reconciliation logic into roots.go
server.go now contains only the lifecycle surface (New, SetLogger, SetToolRegistry, HandleInitialized, HandleRootsChanged). Root reconciliation helpers (loadDesired, snapshotExistingRoots, initializeRoots, replaceRoots, initializeRootsFromSession, listClientRoots, isMethodNotFound, reconcileResult) move into roots.go alongside the existing per-root helpers.

Pure move with one localised cleanup: rename the shadowing roots []*mcp.Root parameter to mcpRoots so the imported roots package stays unambiguous in its new file. No behaviour change.
